requirements.txt
# VeriGuard AI - Production Dependencies
# ========================================

# Core ML/AI
groq>=0.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
torch>=2.0.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# CLI & Formatting
rich>=13.7.0

# Configuration & Utilities
python-dotenv>=1.0.0

# Optional: Cross-encoder reranking
# sentence-transformers handles this, but explicit for clarity

# Development & Testing (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# httpx>=0.25.0
# black>=23.0.0
# ruff>=0.1.0

# Production (optional)
# gunicorn>=21.0.0
# redis>=5.0.0